Step By Step Debug LLM Implemented
Based on the code Aaron wrote, I have integrated it with William's code and made it uniform. I have also created a simple test script that will be built upon to run tests when we are ready to do so. I will be building upon the starter code as time goes on. On top of that I will also add Aaron's tests too since I currently have William's in.